Overhaul availability testing and add expected input languages
Remove the `ai.languageModel.capabilities()` method and its accompanying `AILanguageModelCapabilities` class. Instead, replace it with:

* `ai.languageModel.availability(options)`, which takes the same options as `ai.languageModel.create()`, and returns the corresponding availability.
* `ai.languageModel.params()`, which returns the default and max params (currently top-K and temperature).

Additionally, add the `expectedInputLanguages` option to `create()` and `availability()`. The addition of this option to `create()` allows the web developer to signal the expected input languages ahead of time, allowing the downloading of additional material, or fast-failing if the additional material cannot be supported. The addition of this option to `availability()` replaces the `(await ai.languageModel.capabilities()).languageAvailable()` method.

Closes https://github.com/webmachinelearning/prompt-api/issues/29; see especially https://github.com/webmachinelearning/prompt-api/issues/29#issuecomment-2288105020.

See also https://github.com/webmachinelearning/writing-assistance-apis/pull/22 and https://github.com/webmachinelearning/translation-api/pull/31.